refactor: Relocate export_test_values to new shiny.testmode module
